requirements.txt
absl-py==1.4.0
aiobotocore==2.12.1
aiohttp==3.9.3
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
asn1crypto==1.5.1
atpublic==4.1.0
attrs==23.2.0
blinker==1.7.0
botocore==1.34.51
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
confluent-kafka==2.3.0
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
docker==7.0.0
entrypoints==0.4
filelock==3.13.3
Flask==3.0.2
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.43
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
huggingface-hub==0.22.2
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
Mako==1.3.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
mlflow==2.11.3
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pandas==1.5.3
pillow==10.2.0
platformdirs==3.11.0
protobuf==4.25.3
pyarrow==15.0.2
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytimeparse==1.1.8
pytz==2024.1
pywin32==306
PyYAML==6.0.1
querystring-parser==1.2.4
regex==2023.12.25
requests==2.31.0
retrying==1.3.4
s3fs==2023.12.2
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.12.0
sentence-transformers==2.6.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowflake==0.7.0
snowflake-connector-python==3.7.1
snowflake-ml-python==1.3.1
snowflake-snowpark-python==1.14.0
snowflake._legacy==0.7.0
snowflake.core==0.7.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.4.0
tokenizers==0.15.2
tomlkit==0.12.4
torch==2.2.2
tqdm==4.66.2
transformers==4.39.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.7
waitress==3.0.0
Werkzeug==3.0.2
wrapt==1.16.0
xgboost==1.7.6
yarl==1.9.4
zipp==3.18.1
